AAP-75679: Extract CredentialConfig mixin to eliminate duplicate getter methods
Introduce CredentialConfig mixin in
metrics_utility/automation_controller_billing/package/credential_config.py
with the three methods that were copy-pasted into every CRC-shipping class:
get_ingress_url(), _get_rh_user(), and _get_rh_password().

PackageCRC now inherits from CredentialConfig (listed before base.Package in
the MRO) and drops its own copies of those methods. Behavior is identical —
the same environment variables are read with the same defaults.
